移至 external.ts
